Add TODO to refactor
Some handlers expect certain environment variables, but not all variables are available in all handlers. This could be solved by lazily injecting, but this might result in worse handler performance since the injecting would be done after the Lambda CPU burst period. The better option is to just refactor this module in several submodules so that they can be composed for each handler.

Signed-off-by: rtatton <rdt17@protonmail.com>